// lib/src/types.ts
export interface BaseData {
  id?: string | null;
  _id?: string | null;
  [key: string]: any;
}

export type PathType<RowType extends BaseData> = Exclude<LiteralUnion<keyof RowType, string>, symbol | number>;
// ...
type GetColumnDefinition<RowType extends BaseData, AllDataType extends RowType[] = RowType[]> =
  | WithDataIndex<RowType, AllDataType>
  | WithExactlyOne<RowType, AllDataType, "numerical" | "render" | "colGroup" | "hidden">;

export type ColumnDefinition<RowType extends BaseData, AllDataType extends RowType[] = RowType[]> = GetColumnDefinition<
  RowType,
  AllDataType
>;
// ...
export type ActiveFilters<RowType extends BaseData = any> = ActiveFilter<RowType>[];

export interface OnChangeObject {
  columnFilters: ActiveFilters;
  limit?: number | false;
  skip: number;
  sortDirection: "asc" | "desc" | undefined;
  sortKey: string | null;
}

type AlternateRowColour<RowType extends BaseData> = (data: RowType) => boolean;

export interface RowOptions<RowType extends BaseData> {
  alternateRowColour?: AlternateRowColour<RowType>;
  alternateRowColor?: AlternateRowColour<RowType>;
  rowDisabled?(data: RowType): boolean;
}

export interface Sort {
  key: null | string;
  direction?: "asc" | "desc";
}
// ...
export type TableCellEditHandler<RowType extends BaseData, T = unknown> = (
  update: { path: PathType<RowType>; value: T },
  rowData: RowType,
) => T | Promise<T> | void | Promise<void>;

export interface TableProps<RowType extends BaseData, AllDataType extends RowType[]> {
  /**
   * Custom number of rows after filtering.
   */
  count?: number;
  /**
   * Custom filename for the CSV export file.
   */
  csvFilename?: string;
  /**
   * Disables pagination and shows all rows.
   */
  disablePagination?: boolean;
  /**
   * Enables hiding columns.
   */
  enableHiddenColumns?: boolean;
  /**
   * Enables the export to CSV option.
   */
  exportToCSVOption?: boolean;
  /**
   * A function which is invoked when the table options are changed.
   *
   * Should return the fetched data when it is a CSV export. This is because you may want to return different data if it is during the CSV export.
   * e.g. you may want to export all data as CSV rather than the current page.
   *
   * @param changeObject the object containing the selected filters, pagination and other options.
   * @param isExport whether the function is being invoked during CSV export.
   *
   * @returns the fetched data.
   */
  onChange?: <T extends boolean>(
    changeObject: OnChangeObject,
    isExport: T,
  ) => T extends true ? AllDataType | Promise<AllDataType> : any;
  /**
   * A function invoked when a row is clicked.
   *
   * @param rowData the data of the row.
   * @param e the mouse event.
   */
  rowClick?(rowData: RowType, e: React.MouseEvent<HTMLTableCellElement, MouseEvent>): void;
  /**
   * Options specific to row customisation.
   */
  rowOptions?: RowOptions<RowType>;
  /**
   * The table pagination options.
   */
  rowsPerPageOptions?: TablePaginationProps["rowsPerPageOptions"];
  /**
   * The default number of rows per page.
   */
  rowsPerPageDefault?: number;
  /**
   * The table data.
   */
  tableData: AllDataType;
  /**
   * The MUI table props.
   */
  tableProps?: MUITableProps;
  /**
   * The table's structure.
   */
  tableStructure: ColumnDefinition<RowType, AllDataType>[];
  /**
   * Enables row selection.
   */
  rowsSelectable?: boolean;
  /**
   * A function invoked when a row is selected or deselected.
   *
   * @param rows the selected rows.
   */
  onSelectedRowsChange?(rows: RowType[]): void;
  /**
   * The key to group rows by when selecting a grouped row.
   */
  selectGroupBy?: string;
  /**
   * The default sort options.
   */
  defaultSort?: Sort;
  /**
   * Enables editing rows/cells.
   *
   * @default false when undefined.
   * @default "cells" when true.
   */
  editable?: boolean | "cells" | "rows";
  /**
   * A function invoked when a row/cell is edited.
   *
   * If no function is provided, the row/cell will be updated in place.
   * Otherwise it will expect the function to return the updated value or update the table data itself.
   *
   * @see {@link TableCellEditHandler}
   */
  onEdit?: TableCellEditHandler<RowType>;
  /**
   * Enables the table to be resized.
   */
  resizeable?: boolean;
}